requirements.txt
# python 3.10.6
torch~=1.13.1
torchvision~=0.14.1
numpy~=1.24.2
omegaconf~=2.3.0
hydra-core~=1.3.2
tqdm~=4.65.0
p_tqdm~=1.4.0
matplotlib~=3.7.0
pandas~=1.5.3
scipy~=1.10.1
seaborn~=0.12.2
pillow~=9.4.0
google-cloud-storage~=2.14.0
ipykernel